refactor: rename default date helper
